Add share button, update settings page

+++ app/src/main/java/com/gathr/gathr/ViewGathring.java
@@ -4,51 +4,51 @@ import android.net.Uri;
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBarActivity;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.ListView;
 import android.widget.TextView;
-import org.json.JSONException;
 import org.json.JSONArray;
 import android.content.Intent;
-import android.widget.Toast;
-
-import com.facebook.android.Facebook;
 
 public class ViewGathring extends ActionBarActivity {
 
     private boolean partOf = false;
     private QueryDB DBConn = new QueryDB(this, AuthUser.fb_id, AuthUser.user_id);
-    private String eventId= "1";
-    MyGlobals global = new MyGlobals(this);
-
+    private String eventId = "1";
+    private MyGlobals global = new MyGlobals(this);
+    private boolean loggedin = true;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_view_gathring);
 
-        new SidebarGenerator((DrawerLayout)findViewById(R.id.drawer_layout), (ListView)findViewById(R.id.left_drawer),android.R.layout.simple_list_item_1,this, global.titles, global.links );
-/*
-        Facebook fb = new Facebook(R.string.facebook_app_id);
-        ShareLinkContent content = new ShareLinkContent.Builder()
-                .setContentUrl(Uri.parse("https://developers.facebook.com"))
-                .build();
-        ShareButton shareButton = (ShareButton)findViewById(R.id.fb_share_button);
-        shareButton.setShareContent(content);
-*/
         try{
-            String event_organizer;
-            Bundle extras = getIntent().getExtras();
+            Intent intent = getIntent();
+            Bundle extras = intent.getExtras();
             if(extras != null)
                 eventId = (String)extras.get("eventId");
 
-            String result;
+            Uri data = intent.getData();
+            if(data != null) {
+                String[] url = data.toString().split("/");
+                eventId = url[(url.length - 1)];
+            }
 
-            DBConn.executeQuery("SELECT * FROM EVENTS WHERE Id =" + eventId);
-            result = DBConn.getResults();
+            if(AuthUser.user_id != null){
+                new SidebarGenerator((DrawerLayout)findViewById(R.id.drawer_layout), (ListView)findViewById(R.id.left_drawer),android.R.layout.simple_list_item_1,this, global.titles, global.links );
+            }else{
+                String[] title = {"Login"};
+                Class<?>[] link = {MainActivity.class};
+                new SidebarGenerator((DrawerLayout)findViewById(R.id.drawer_layout), (ListView)findViewById(R.id.left_drawer),android.R.layout.simple_list_item_1,this, title, link );
+                loggedin = false;
+            }
+
+            QueryDB getEvent = new QueryDB(this,"getEvent.php",true);
+            getEvent.executeQuery(eventId);
+            String result = getEvent.getResults();
 
             JSONArray json;
             json = new JSONArray(result);
@@ -59,7 +59,7 @@ public class ViewGathring extends ActionBarActivity {
             String state = json.getJSONObject(0).getString("State");
             String time = json.getJSONObject(0).getString("Time");
             String capacity = json.getJSONObject(0).getString("Capacity");
-            event_organizer = json.getJSONObject(0).getString("Organizer").trim();
+            String event_organizer = json.getJSONObject(0).getString("Organizer").trim();
 
             ((TextView)findViewById(R.id.gathring_name_text)).setText(eventName);
             ((TextView)findViewById(R.id.gathring_description_text)).setText(description);
@@ -70,12 +70,13 @@ public class ViewGathring extends ActionBarActivity {
             ((TextView)findViewById(R.id.gathring_time_text)).setText(global.normalTime(time));
 
             TextView buttonText = (TextView) findViewById(R.id.join_leave_button);
-
-            if(!event_organizer.equals(AuthUser.user_id)) {
+            if(!loggedin){
+                buttonText.setText("Login");
+            }else if(!event_organizer.equals(AuthUser.user_id)) {
                 DBConn.executeQuery("SELECT COUNT(User_Id) AS Count FROM JOINED_EVENTS WHERE User_Id = " + AuthUser.user_id + " AND Event_Id = " + eventId + ";");
                 result = DBConn.getResults();
                 json = new JSONArray(result);
-                String count = json.getJSONObject(json.length() - 1).getString("Count").trim();
+                String count = json.getJSONObject(0).getString("Count").trim();
 
                 if (count.equals("0")) {
                     buttonText.setText("Join");
@@ -95,21 +96,26 @@ public class ViewGathring extends ActionBarActivity {
     public void joinOrLeave(View view){
         try {
             TextView buttonText = (TextView) findViewById(R.id.join_leave_button);
-            if(!partOf) {
-                DBConn.executeQuery("INSERT INTO JOINED_EVENTS(User_Id, Event_Id) VALUES (" + AuthUser.user_id + "," + eventId + ");");
-                //DBConn.getResults(); // We do not need to wait
-                global.tip("Welcome to the Gathring");
-                partOf = true;
-                buttonText.setText("Leave");
-            }else{
-                DBConn.executeQuery("DELETE FROM JOINED_EVENTS WHERE User_Id=" + AuthUser.user_id + " and Event_Id= " + eventId + ";");
-                //DBConn.getResults(); // We do not need to wait
-                global.tip("You have left the Gathring");
-                partOf = false;
-                buttonText.setText("Join");
+            if(!loggedin){
+                Intent intent = new Intent(this, MainActivity.class); // Send them back to login page
+                startActivity(intent);
+            }else {
+                if (!partOf) {
+                    DBConn.executeQuery("INSERT INTO JOINED_EVENTS(User_Id, Event_Id) VALUES (" + AuthUser.user_id + "," + eventId + ");");
+                    //DBConn.getResults(); // We do not need to wait
+                    global.tip("Welcome to the Gathring");
+                    partOf = true;
+                    buttonText.setText("Leave");
+                } else {
+                    DBConn.executeQuery("DELETE FROM JOINED_EVENTS WHERE User_Id=" + AuthUser.user_id + " and Event_Id= " + eventId + ";");
+                    //DBConn.getResults(); // We do not need to wait
+                    global.tip("You have left the Gathring");
+                    partOf = false;
+                    buttonText.setText("Join");
+                }
             }
-        }catch(GathrException e){
-            global.errorHandler (e);
+        }catch(Exception e){
+            global.errorHandler(e);
         }
     }
 
@@ -128,7 +134,13 @@ public class ViewGathring extends ActionBarActivity {
         int id = item.getItemId();
 
         //noinspection SimplifiableIfStatement
-        if (id == R.id.action_settings) {
+        if (id == R.id.action_share) {
+            Intent shareIntent = new Intent();
+            shareIntent.setAction(Intent.ACTION_SEND);
+            shareIntent.setType("text/plain");
+            shareIntent.putExtra(Intent.EXTRA_TEXT, "Check out this event on Gathr: http://www.gathr.com/viewEvent/"+ eventId +" !" );
+            startActivity(Intent.createChooser(shareIntent, "Share this event"));
+
             return true;
         }
